\name{FileServer}
\alias{FileServer}
\title{Start file server}
\usage{
  FileServer(config_file, title = "File Server",
    fun_label = "", button_label = "Run", ...)
}
\arguments{
  \item{config_file}{config file path}

  \item{title}{Page title}

  \item{fun_label}{Function's select input label}

  \item{button_label}{Download button label}

  \item{...}{see runApp()}
}
\description{
  ...
}

